Fix Doris OVER-clause visitor scope, BITOR/ARRAY_LAST AST shape, and window frame text
