+++ test/it/parser/src/main/java/org/apache/shardingsphere/test/it/sql/parser/internal/asserts/segment/window/WindowClauseAssert.java
@@ -27,6 +27,7 @@ import org.apache.shardingsphere.sql.parser.statement.core.segment.generic.Windo
 import org.apache.shardingsphere.test.it.sql.parser.internal.asserts.SQLCaseAssertContext;
 import org.apache.shardingsphere.test.it.sql.parser.internal.asserts.segment.SQLSegmentAssert;
 import org.apache.shardingsphere.test.it.sql.parser.internal.asserts.segment.column.ColumnAssert;
+import org.apache.shardingsphere.test.it.sql.parser.internal.asserts.segment.expression.ExpressionAssert;
 import org.apache.shardingsphere.test.it.sql.parser.internal.asserts.segment.orderby.OrderByClauseAssert;
 import org.apache.shardingsphere.test.it.sql.parser.internal.cases.parser.jaxb.segment.impl.column.ExpectedColumn;
 import org.apache.shardingsphere.test.it.sql.parser.internal.cases.parser.jaxb.segment.impl.orderby.ExpectedOrderByClause;
@@ -72,6 +73,7 @@ public final class WindowClauseAssert {
     public static void assertIs(final SQLCaseAssertContext assertContext, final WindowItemSegment actual, final ExpectedWindowItem expected) {
         SQLSegmentAssert.assertIs(assertContext, actual, expected);
         assertPartitionBy(assertContext, actual, expected);
+        assertFrame(assertContext, actual, expected);
         ExpectedOrderByClause expectedOrderByClause = expected.getOrderByClause();
         OrderBySegment actualOrderBySegment = actual.getOrderBySegment();
         if (null == expectedOrderByClause) {
@@ -94,5 +96,12 @@ public final class WindowClauseAssert {
             assertThat(assertContext.getText("Window item partition by expression assertion error: "), actualExpression, isA(ColumnSegment.class));
             ColumnAssert.assertIs(assertContext, (ColumnSegment) actualExpression, each);
         }
+    }
+    
+    private static void assertFrame(final SQLCaseAssertContext assertContext, final WindowItemSegment actual, final ExpectedWindowItem expected) {
+        if (null == expected.getFrameClause()) {
+            return;
+        }
+        ExpressionAssert.assertExpression(assertContext, actual.getFrameClause(), expected.getFrameClause());
     }
 }
