+++ test/it/parser/src/main/resources/case/ddl/create-view.xml
@@ -130,6 +130,24 @@
         </select>
     </create-view>
     
+    <create-view sql-case-id="create_view_with_window_doris" view-definition="SELECT MAX(order_id) OVER (PARTITION BY user_id) FROM t_order">
+        <view name="v" start-index="12" stop-index="12" />
+        <select>
+            <projections start-index="24" stop-index="64">
+                <aggregation-projection type="MAX" expression="MAX(order_id) OVER (PARTITION BY user_id)" start-index="24" stop-index="64">
+                    <window-item start-index="38" stop-index="64">
+                        <partition-by>
+                            <column name="user_id" start-index="57" stop-index="63" />
+                        </partition-by>
+                    </window-item>
+                </aggregation-projection>
+            </projections>
+            <from>
+                <simple-table name="t_order" start-index="71" stop-index="77" />
+            </from>
+        </select>
+    </create-view>
+
     <create-view sql-case-id="create_view_with_check_option" view-definition="SELECT * FROM comedies WHERE classification = 'U'">
         <view name="universal_comedies" start-index="12" stop-index="29" />
         <select>
