+++ test/it/parser/src/main/resources/case/dml/select-window.xml
@@ -152,9 +152,15 @@
         <projections start-index="7" stop-index="86">
             <aggregation-projection type="SUM" expression="SUM(v) OVER (PARTITION BY k ORDER BY v ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING)" start-index="7" stop-index="86">
                 <window-item start-index="14" stop-index="86">
+                    <partition-by>
+                        <column name="k" start-index="33" stop-index="33" />
+                    </partition-by>
                     <order-by start-index="35" stop-index="44">
                         <column-item name="v" order-direction="ASC" start-index="44" stop-index="44" />
                     </order-by>
+                    <frame>
+                        <common-expression text="ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING" start-index="46" stop-index="85" />
+                    </frame>
                 </window-item>
                 <parameter>
                     <column name="v" start-index="11" stop-index="11" />
