+++ test/it/parser/src/main/resources/sql/supported/dml/select-special-function.xml
@@ -291,10 +291,13 @@
     <sql-case id="select_st_convexhull" value="SELECT ST_ConvexHull(ST_GeomFromText('MULTIPOINT(5 0,25 0,15 10,15 25)'))" db-types="MySQL,Doris" />
     <sql-case id="select_st_crosses" value="SELECT ST_Crosses(ST_GeomFromText('POINT(1 1)'), ST_GeomFromText('POINT(2 2)'))" db-types="MySQL,Doris" />
     <sql-case id="select_bitor" value="SELECT BITOR(3,5)" db-types="Doris" />
+    <sql-case id="select_bitor_with_columns" value="SELECT BITOR(user_id, order_id + 1) FROM t_order" db-types="Doris" />
+    <sql-case id="select_bitor_with_parameter_marker" value="SELECT BITOR(?, ?)" db-types="Doris" />
     <sql-case id="select_bitxor" value="SELECT BITXOR(3,5)" db-types="Doris" />
     <sql-case id="select_bitmap_to_string" value="SELECT BITMAP_TO_STRING(TO_BITMAP(1))" db-types="Doris" />
     <sql-case id="select_bitmap_to_string_multiple" value="SELECT BITMAP_TO_STRING(BITMAP_OR(TO_BITMAP(1), TO_BITMAP(2)))" db-types="Doris" />
     <sql-case id="select_array_last" value="SELECT ARRAY_LAST(x -&gt; x &gt; 0, [1, 2, 3, -1, 2])" db-types="Doris" />
+    <sql-case id="select_array_last_with_multiple_parameters" value="SELECT ARRAY_LAST((x, y) -&gt; x &gt; y, [1, 2, 3, 4, 5], [1.1, 2.2, 3.3, 4.4, 5.5])" db-types="Doris" />
     <sql-case id="select_window_max_distinct_doris" value="SELECT MAX(DISTINCT order_id) OVER (PARTITION BY user_id) AS max_order_id FROM t_order" db-types="Doris" />
     <sql-case id="select_window_max_all_doris" value="SELECT MAX(ALL order_id) OVER (PARTITION BY user_id) AS max_order_id FROM t_order" db-types="Doris" />
     <sql-case id="select_instr" value="SELECT INSTR('foobar','bar')" db-types="Doris" />
